Werk Work - After some playing around there were issues with auto_incremented IDs. UUIDs now have replaced all integer IDs for personas, which are assigned upon creation. - SQLite compatibility. Running well! SQLite is highly functional. - Continued fixing of Errors.

+++ src/main/java/net/lordofthecraft/arche/save/PersonaField.java
@@ -29,7 +29,7 @@ public enum PersonaField {
     ENDERINV("ender_inv", PersonaTable.VITALS, JDBCType.VARCHAR),
     HEALTH("health", PersonaTable.VITALS, JDBCType.DOUBLE),
     FOOD("food", PersonaTable.VITALS, JDBCType.INTEGER),
-    SATURATION("saturation", PersonaTable.VITALS, JDBCType.INTEGER);
+    SATURATION("saturation", PersonaTable.VITALS, JDBCType.FLOAT);
 
     private static final String STATEMENT_SUFFIX = " = ? WHERE persona_id";
